Set-up base files from using dotnet instruction
This commit sets-up the base files runing the next commands:
1.- 'dotnet new sln -n SocialApiCs', to start a new solution
2.- 'dotnet new webapi -n SocialApiCs.Api' to add our api project
3.- 'dotnet new classlib -n SocialApiCs.core' to add our core project
4.- 'dotnet new classlib -n SocialApiCs.Infrastructure'

I added a folder for future testing and two more project inside to
performe unit tests and integration tests. In addition references were
made from the core and infrastructure to the api project and from the
core to the infrastructure project.